+++ packages/akashic-cli-serve/src/client/common/types/ScenarioEventData.ts
@@ -1,0 +1,11 @@
+type CommandName = "screenshot" | "finish";
+
+export interface ScenarioEventData {
+	type: "scenario";
+	command: {
+		name: CommandName;
+		options?: {
+			fileName: string;
+		};
+	};
+}